refactor: code style, types
